pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "h3c-sgbox-translator"
version = "2.0.0"
description = "H3C firewall syslog to SGBox SIEM translator — async, Debian 13 ready"
readme = "docs/BUG_AUDIT_REPORT.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "H3C Translator Team"},
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: No Input/Output (Daemon)",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: System :: Logging",
    "Topic :: System :: Networking :: Firewalls",
    "Framework :: AsyncIO",
]
dependencies = [
    "aiohttp>=3.13,<4.0",
    "uvloop>=0.22",
    "structlog>=25.1",
    "tenacity>=9.0",
    "pyOpenSSL>=25.0",
    "python-gnupg>=0.5.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-aiohttp>=1.0",
]

[project.scripts]
h3c-translator = "src.translator:main"

[tool.setuptools.packages.find]
include = ["src*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
